requirements.txt
allure-pytest==2.15.3
certifi==2025.8.3
charset-normalizer==3.4.3
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
packaging==25.0
playwright==1.55.0
pluggy==1.6.0
pyee==13.0.0
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-slugify==8.0.4
requests==2.32.5
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.5.0
